backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
asyncio==3.4.3
aiofiles==23.2.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8
# QR Code and data processing
qrcode==8.2
pyzbar==0.1.9
pillow==11.2.1
# Text processing and regex
regex==2023.12.25
# AI integration for categorization
ollama==0.1.7
# Data validation and cleaning
validators==0.22.0
# Date parsing
python-dateutil==2.8.2 